Archlinux: schroot-hybris
Move from schroot-android to schroot-hybris

Add an unified patch to support Android environment setting.
That permit to set LD_LIBRARY_PATH, PATH, PS1, ANDROID_ROOT, LD_PRELOAD into the chroot.
The home of the user is always set to / as Android is not compliant with LHS to
avoid directory change issue.

So it is now possible to run a command directly.
eg: schroot -r -c droid -- ls -l /

Packaging for hybris-device doesn't need anymore to use a static busybox for sh and
to create a root folder with profile that export environment. Of course you need to
set 'android-environment=true' into the schroot configuration.